extract doi construction into its own file
